Modularise the links at bottom of page
The links are used on both homepage and about pages so it makes sense to have them as a module (mixin) in pug. At a later point, may want to decide if these will be on all pages in which case they should be worked into one of the templates.