Implement email sending form on back end
Uses Nodemailer to send an email to ourselves with the query/comment/whatever that was typed into the form on the email. Should be pretty hard to programatically hit this route as captcha is required.

To do: Validate that user has entered information into all fields before allowing them to send. Implement on front and back end.

Note: For security, sensitive information has been moved to separate JSON file. Make sure to upload this file to server when deploying.